fix: 잘못된 룰 수정, 타입스크립트 다운그레이드

+++ rules/typescript.js
@@ -48,7 +48,7 @@ module.exports = {
     '@typescript-eslint/prefer-namespace-keyword': 'error',
     '@typescript-eslint/quotes': [
       'error',
-      'double',
+      'single',
       {'allowTemplateLiterals': true }
     ],
     '@typescript-eslint/triple-slash-reference': 'error',
